requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.0.74
scipy==1.10.1
matplotlib==3.7.1
fastapi==0.100.0
uvicorn==0.23.2
pydantic==2.0.3
gym==0.26.2
stable-baselines3==2.0.0
pyyaml==6.0
astropy==5.3.2
scikit-learn==1.3.0
pillow==10.0.0